Move time increment hook up in DOM
